+++ src/sdk/SessionKeyValidator/interfaces.ts
@@ -4,7 +4,21 @@ export interface GenerateSessionKeyResponse {
     oldSessionKey?: string;
 }
 
-export interface EnableSessionKeyResponse {
+export interface GetSessionKeyResponse {
+    sessionKey: string;
+}
+
+export interface DeleteSessionKeyResponse {
+    account: string;
+    chainId: number;
+    message: string;
+}
+
+export interface SessionKeyResponse {
     userOpHash: string;
     sessionKey: string;
+}
+
+export interface GetNonceResponse {
+    nonce: number;
 }